requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
fastapi[all]
redis
requests
httpx
asyncio
pytest